Fixing a bug relating to the processing of the file names entered in the standard input in main.

+++ segment_data.py
@@ -19,7 +19,7 @@ class SegmentData:
         """
             This method loads a segment from a file.
         """
-        self.data=loadtxt(file_name)
+        self.data=loadtxt(file_name, dtype='int')
 
 
     def load_string(self, input_string):
